+++ ojdbc-provider-jackson-oson/src/main/java/oracle/jdbc/provider/oson/OsonFactory.java
@@ -1,4 +1,4 @@
-package oson;
+package oracle.jdbc.provider.oson;
 
 import com.fasterxml.jackson.core.JsonEncoding;
 import com.fasterxml.jackson.core.JsonFactory;
