+++ ojdbc-provider-jackson-oson/src/main/java/oracle/jdbc/provider/oson/deser/OsonBigIntegerDeserializer.java
@@ -1,4 +1,4 @@
-package oson.deser;
+package oracle.jdbc.provider.oson.deser;
 
 import com.fasterxml.jackson.core.JacksonException;
 import com.fasterxml.jackson.core.JsonParser;
