+++ ojdbc-provider-jackson-oson/src/main/java/oracle/jdbc/provider/oson/provider/JacksonOsonConverter.java
@@ -1,4 +1,4 @@
-package oson.provider;
+package oracle.jdbc.provider.oson.provider;
 
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.ObjectMapper;
