+++ ojdbc-provider-jackson-oson/src/main/java/oracle/jdbc/provider/oson/provider/JacksonOsonProvider.java
@@ -1,4 +1,4 @@
-package oson.provider;
+package oracle.jdbc.provider.oson.provider;
 
 import oracle.jdbc.spi.JsonProvider;
 import oracle.jdbc.spi.OsonConverter;
